(#81) Default to accepting file requests w/ unknown extensions
This won't allow file extensions which IIS is expressly blocking, but in
the case where the server was defaulted to require file extensions to be
explicitly permitted, this overrides it. (i.e. This allows requests for .css,
.js, txt, .html, etc.)

Since print requests will contain potentially non-ASCII characters (and
as we're copying this entry from wcms-cde-config with experiences
learned there), we'll also allow high-bit characters.